worker/requirements.txt
GitPython==3.1.42
pika==1.3.2
semgrep==1.66.0